[SPARK-43879][CONNECT] Decouple handle command and send response on server side
